Breakout types into own files, rename
