+++ mosaic-terminal/src/nativeMain/kotlin/com/jakewharton/mosaic/terminal/Tty.kt
@@ -4,7 +4,6 @@ import com.jakewharton.mosaic.terminal.event.Event
 import kotlinx.cinterop.COpaquePointer
 import kotlinx.cinterop.CPointer
 import kotlinx.cinterop.StableRef
-import kotlinx.cinterop.addressOf
 import kotlinx.cinterop.alloc
 import kotlinx.cinterop.asStableRef
 import kotlinx.cinterop.free
@@ -12,7 +11,6 @@ import kotlinx.cinterop.nativeHeap
 import kotlinx.cinterop.ptr
 import kotlinx.cinterop.staticCFunction
 import kotlinx.cinterop.useContents
-import kotlinx.cinterop.usePinned
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.channels.Channel.Factory.UNLIMITED
 
@@ -62,7 +60,7 @@ public actual object Tty {
 		return TerminalReader(reader, events, emitDebugEvents)
 	}
 
-	internal actual fun stdinWriter(emitDebugEvents: Boolean): StdinWriter {
+	internal actual fun platformInputWriter(): PlatformInputWriter {
 		val events = Channel<Event>(UNLIMITED)
 
 		// TODO Fix all this duplication, ownership
@@ -88,99 +86,11 @@ public actual object Tty {
 
 		val readerPtr = stdinWriter_getReader(writerPtr)!!
 		val platformInput = PlatformInput(readerPtr, handlerPtr, handlerRef)
-		val terminalReader = TerminalReader(platformInput, events, emitDebugEvents)
-		return StdinWriter(writerPtr, terminalReader)
+		return PlatformInputWriter(writerPtr, events, platformInput)
 	}
 
 	internal fun throwError(error: UInt): Nothing {
 		throw RuntimeException(error.toString())
-	}
-}
-
-internal actual class PlatformInput internal constructor(
-	ptr: CPointer<stdinReader>,
-	private val handlerPtr: CPointer<platformEventHandler>?,
-	private val handlerRef: StableRef<PlatformEventHandler>?,
-) : AutoCloseable {
-	private var ptr: CPointer<stdinReader>? = ptr
-
-	actual fun read(buffer: ByteArray, offset: Int, count: Int): Int {
-		buffer.usePinned {
-			stdinReader_read(ptr, it.addressOf(offset), count).useContents {
-				if (error == 0U) return this.count
-				Tty.throwError(error)
-			}
-		}
-	}
-
-	actual fun readWithTimeout(buffer: ByteArray, offset: Int, count: Int, timeoutMillis: Int): Int {
-		buffer.usePinned {
-			stdinReader_readWithTimeout(ptr, it.addressOf(offset), count, timeoutMillis).useContents {
-				if (error == 0U) return this.count
-				Tty.throwError(error)
-			}
-		}
-	}
-
-	actual fun interrupt() {
-		val error = stdinReader_interrupt(ptr)
-		if (error == 0U) return
-		Tty.throwError(error)
-	}
-
-	actual override fun close() {
-		ptr?.let { ptr ->
-			this.ptr = null
-
-			val error = stdinReader_free(ptr)
-			handlerPtr?.let(nativeHeap::free)
-			handlerRef?.dispose()
-
-			if (error == 0U) return
-			Tty.throwError(error)
-		}
-	}
-}
-
-internal actual class StdinWriter internal constructor(
-	private var ptr: CPointer<stdinWriter>?,
-	actual val reader: TerminalReader,
-) : AutoCloseable {
-	actual fun write(buffer: ByteArray) {
-		val error = buffer.usePinned {
-			stdinWriter_write(ptr, it.addressOf(0), buffer.size)
-		}
-		if (error == 0U) return
-		Tty.throwError(error)
-	}
-
-	actual fun focusEvent(focused: Boolean) {
-		stdinWriter_focusEvent(ptr, focused)
-	}
-
-	actual fun keyEvent() {
-		stdinWriter_keyEvent(ptr)
-	}
-
-	actual fun mouseEvent() {
-		stdinWriter_mouseEvent(ptr)
-	}
-
-	actual fun resizeEvent(columns: Int, rows: Int, width: Int, height: Int) {
-		stdinWriter_resizeEvent(ptr, columns, rows, width, height)
-	}
-
-	actual override fun close() {
-		ptr?.let { ref ->
-			this.ptr = null
-
-			reader.close()
-
-			val error = stdinWriter_free(ref)
-
-			if (error == 0U) return
-			Tty.throwError(error)
-		}
 	}
 }
